pyproject.toml
[tool.poetry]
name = "trademaster"
version = "0.1.0"
description = ""
authors = ["Rodrigo Carvalho <rodrigofmcarvalho@gmail.com>"]
readme = "README.md"
packages = [{include = "trademaster"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
psycopg2-binary = "^2.9.9"
faker = "^19.12.0"
fastapi = "^0.104.1"
kafka-python = "^2.0.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"
pydantic = {extras = ["email"], version = "^2.4.2"}
pandas = "^2.1.2"
boto3 = "^1.28.75"
pyarrow = "^14.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
